feat(api): Implement hiding of objects
Use new appendUser middleware to send the logged-in-user/undefined to API controller index/show functions. For index functions, remove hidden items if the user is not admin. For show functions, return 401 error if the item is hidden and the user is not admin. Update DB models to use 'hidden' field. The following objects can be hidden: file, gallery, photo, project, post.